requirements.txt
Django
psycopg2-binary
dj-database-url
whitenoise
python-dotenv
requests
groq